mnist/requirements.txt
Flask==2.1.2
numpy==1.22.4
tensorflow-macos==2.9.2